Move options to common folder.
