+++ extensions/mssql/src/webviews/pages/TableExplorer/TableDataGrid.tsx
@@ -29,7 +29,7 @@ import TableExplorerCustomPager from "./TableExplorerCustomPager";
 import { slickGridLocales } from "./commonGridOptions";
 import "@slickgrid-universal/common/dist/styles/css/slickgrid-theme-fluent.css";
 import "./TableDataGrid.css";
-import { baseFluentGridOption } from "../base-fluent-grid-options";
+import { baseFluentGridOption } from "../../common/FluentSlickGrid/base-fluent-grid-options";
 
 interface TableDataGridProps {
     resultSet: EditSubsetResult | undefined;
